+++ .eslintrc.js
@@ -13,5 +13,7 @@ module.exports = {
     ecmaVersion: 2018,
   },
   rules: {
+    'class-methods-use-this': 'off',
+    'no-console': 'off',
   },
 };
