+++ app/kairos-broker.js
@@ -1,72 +1,38 @@
 /* eslint-disable */
 const Redis = require('ioredis');
-const assert = require('assert');
 
 const Log = require('./log.js');
 const Colors = require('./colors.js');
 const DateTime = require('./date-time.js');
+const Config = require('../config/config.js');
+const RedisConnections = require('./redis-connections');
+const RedisControl = require('./redis-control');
 
 const dateTime = new DateTime();
 const log = new Log(dateTime, Colors);
 
-function assertBrokerOptions(brokerOptions) {
-  assert(brokerOptions, '"brokerOptions" is required to create a Redis client with sc-redis');
-  assert(brokerOptions.host, '"brokerOptions.host" is required to create a Redis client with sc-redis');
-  assert(brokerOptions.port, '"brokerOptions.port" is required to create a Redis client with sc-redis');
-}
-
-function throwMissingRedisClientError(clientName) {
-  throw new Error('Missing "' + clientName + '" option. Both "pubClient" and "subClient" must be specified if passing your own clients.');
-}
-
 module.exports.attach = function (broker) {
 
-  const instanceId = broker.instanceId;
-  var subClient = null;
-  var pubClient = null;
-
-  if (!subClient && !pubClient) {
-    const brokerOptions = broker.options.brokerOptions;
-    assertBrokerOptions(brokerOptions);
-
-    try {
-      subClient = new Redis({
-        host: brokerOptions.host,
-        port: brokerOptions.port,
-        autoResubscribe: false,
-        retryStrategy: () => 5000,
-      });
-      pubClient = new Redis({
-        host: brokerOptions.host,
-        port: brokerOptions.port,
-        autoResubscribe: false,
-        retryStrategy: () => 5000,
-      });
-    } catch (error) {
-    }
-  } else if (!subClient && pubClient) {
-    throwMissingRedisClientError("subClient");
-  } else if (subClient && !pubClient) {
-    throwMissingRedisClientError("pubClient");
+  function bindSubscribe(connections) {
+    log.debug('yellow', `Broker PID: ${process.pid} - REDIS Binding subscribe`);
+    return connections[0].subClient.subscribe.bind(connections[0].subClient);
   }
 
-  subClient.on('ready', () => {
-    log.debug('green', `Broker PID: ${process.pid} - REDIS Connected at ${subClient.options.host} :${subClient.options.port}`);
-  });
+  function bindUnsubscribe(connections) {
+    log.debug('yellow', `Broker PID: ${process.pid} - REDIS Binding unsubscribe`);
+    return connections[0].subClient.unsubscribe.bind(connections[0].subClient);
+  }
 
-  subClient.on('reconnecting', () => {
-    log.debug('red', `Broker PID: ${process.pid} - REDIS Lost Connection with ${subClient.options.host} :${subClient.options.port} retrying in 5 seconds`);
-  });
+  const instanceId = broker.instanceId;
 
-  subClient.on('error', () => {
-  });
+  const redisConnections = new RedisConnections(Config, Redis);
+  const connections = redisConnections.connect();
 
-  pubClient.on('error', () => {
-  });
+  const redisControl = new RedisControl(log, connections, process.pid);
+  redisControl.listen();
 
-  broker.on('subscribe', subClient.subscribe.bind(subClient));
-
-  broker.on('unsubscribe', subClient.unsubscribe.bind(subClient));
+  broker.on('subscribe', bindSubscribe(connections));
+  broker.on('unsubscribe', bindUnsubscribe(connections));
 
   broker.on('publish', (channel, data) => {
     if (data instanceof Object) {
@@ -83,12 +49,12 @@ module.exports.attach = function (broker) {
       data = instanceId + data;
     }
 
-    pubClient.publish(channel, data);
+    connections[0].pubClient.publish(channel, data);
   });
 
   var instanceIdRegex = /^[^\/]*\//;
 
-  subClient.on('message', (channel, message) => {
+  connections[0].subClient.on('message', (channel, message) => {
     var sender = null;
     message = message.replace(instanceIdRegex, function (match) {
       sender = match.slice(0, -1);
@@ -110,4 +76,4 @@ module.exports.attach = function (broker) {
       broker.publish(channel, data);
     }
   });
-};
+}
