+++ config/config.js
@@ -1,10 +1,10 @@
 module.exports = [
-    {
-        host : 'localhost',
-        port : '6379',
-    },
-    {
-        host : 'localhost',
-        port : '6380',
-    }
-]
+  {
+    host: 'localhost',
+    port: 6379,
+  },
+  {
+    host: 'localhost',
+    port: 6380,
+  },
+];
