+++ public/subscribe.html
@@ -3,7 +3,7 @@
   <head>
     <title>Socket Client</title>
     <link rel="icon" type="image/png" href="/favicon.ico">
-    <script type="text/javascript" src="/socketcluster.js"></script>
+    <script type="text/javascript" src="/kairos.js"></script>
   </head>
     <body>
     <input id="channel" />
