+++ worker.js
@@ -5,14 +5,13 @@ const serveStatic = require('serve-static');
 const path = require('path');
 const healthChecker = require('sc-framework-health-check');
 
-const Redis = require('ioredis');
-const Config = require('./config/config.js');
 const Log = require('./app/log.js');
-const RedisPubSub = require('./app/redis-pubsub.js');
-const RedisConnections = require('./app/redis-connections.js');
 const Colors = require('./app/colors.js');
 const DateTime = require('./app/date-time.js');
-const SocketControl = require('./app/socket-control.js');
+const WorkerControl = require('./app/worker-control');
+
+const dateTime = new DateTime();
+const log = new Log(dateTime, Colors);
 
 class Worker extends SCWorker {
   run() {
@@ -24,17 +23,10 @@ class Worker extends SCWorker {
     healthChecker.attach(this, app);
     httpServer.on('request', app);
 
-    const dateTime = new DateTime();
-    const log = new Log(dateTime, Colors);
-    const redisConnections = new RedisConnections(Config, Redis);
-    const connections = redisConnections.connect();
-    const socketControl = new SocketControl(log, scServer);
+    const workerControl = new WorkerControl(log, scServer, process.pid);
+    workerControl.listen();
 
-    log.debug('white', '   >> Worker PID:', process.pid);
-
-    const pubSub = new RedisPubSub(connections, log, socketControl);
-    pubSub.monitoring();
-    socketControl.listen();
+    log.debug('cyan', `Worker PID: ${process.pid} - Starting`);
   }
 }
